+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/util/AuthorUtil.java
@@ -1,0 +1,22 @@
+package uk.nhs.adaptors.pss.translator.util;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.hl7.fhir.dstu3.model.Reference;
+import org.hl7.v3.RCMRMT030101UK04Author;
+import org.hl7.v3.RCMRMT030101UK04EhrComposition;
+
+import java.util.Optional;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class AuthorUtil {
+
+    private static final String PRACTITIONER_REFERENCE_PREFIX = "Practitioner/%s";
+
+    public static Optional<Reference> getAuthorReference(RCMRMT030101UK04EhrComposition ehrComposition) {
+
+        Optional<RCMRMT030101UK04Author> author = Optional.ofNullable(ehrComposition.getAuthor());
+        return author.map(a -> new Reference(PRACTITIONER_REFERENCE_PREFIX.formatted(a.getAgentRef().getId().getRoot())));
+    }
+
+}