+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/util/ParticipantReferenceUtil.java
@@ -2,8 +2,9 @@ package uk.nhs.adaptors.pss.translator.util;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 import org.hl7.fhir.dstu3.model.Reference;
 import org.hl7.v3.II;
 import org.hl7.v3.RCMRMT030101UK04AgentRef;
@@ -11,6 +12,7 @@ import org.hl7.v3.RCMRMT030101UK04EhrComposition;
 import org.hl7.v3.RCMRMT030101UK04Participant;
 import org.hl7.v3.RCMRMT030101UK04Participant2;
 
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ParticipantReferenceUtil {
     private static final String PRF_TYPE_CODE = "PRF";
     private static final String PPRF_TYPE_CODE = "PPRF";
@@ -20,7 +22,7 @@ public class ParticipantReferenceUtil {
         RCMRMT030101UK04EhrComposition ehrComposition) {
         var nonNullFlavorParticipants = participantList.stream()
             .filter(ParticipantReferenceUtil::isNotNullFlavour)
-            .collect(Collectors.toList());
+            .toList();
 
         var pprfParticipants = getParticipantReference(nonNullFlavorParticipants, PPRF_TYPE_CODE);
         if (pprfParticipants.isPresent()) {
@@ -41,6 +43,7 @@ public class ParticipantReferenceUtil {
     }
 
     private static Optional<String> getParticipantReference(List<RCMRMT030101UK04Participant> participantList, String typeCode) {
+
         return participantList.stream()
             .filter(participant -> hasTypeCode(participant, typeCode))
             .filter(ParticipantReferenceUtil::hasAgentReference)
@@ -49,6 +52,23 @@ public class ParticipantReferenceUtil {
             .filter(II::hasRoot)
             .map(II::getRoot)
             .findFirst();
+    }
+
+    public static Reference getParticipant2Reference(RCMRMT030101UK04EhrComposition ehrComposition, String typeCode) {
+
+        var participant2Reference = ehrComposition.getParticipant2().stream()
+            .filter(participant2 -> participant2.getNullFlavor() == null)
+            .filter(participant2 -> typeCode.equals(participant2.getTypeCode().get(0)))
+            .map(RCMRMT030101UK04Participant2::getAgentRef)
+            .map(RCMRMT030101UK04AgentRef::getId)
+            .filter(II::hasRoot)
+            .map(II::getRoot)
+            .findFirst();
+
+        if (participant2Reference.isPresent()) {
+            return new Reference(PRACTITIONER_REFERENCE_PREFIX.formatted(participant2Reference.get()));
+        }
+        return null;
     }
 
     private static Optional<String> getParticipant2Reference(RCMRMT030101UK04EhrComposition ehrComposition) {