+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 * Fixed issue with some blood pressure comments not being correctly mapped.
+* Fixed issue with practitioner not being correctly populated.
 
 ## [0.14] - 2023-10-17